requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==3.1.0
dataclasses==0.6
dataclasses-serialization==1.3.1
Django==4.1.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==1.1.1
more-properties==1.1.1
multidict==6.0.4
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
toposort==1.7
typing-inspect==0.8.0
typing_extensions==4.4.0
websockets==10.3
yarl==1.8.2
